Revert "chore: Fix IDs for some sidebar items"
This reverts commit a267c0ee7d0fb159977e8a5963e22e0b57c73f46.
