+++ packages/core/src/renderer/components/workloads-daemonsets/daemonsets-sidebar-item.injectable.tsx
@@ -4,29 +4,29 @@
  */
 import { getInjectable } from "@ogre-tools/injectable";
 
-import statefulsetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/statefulsets/statefulsets-route.injectable";
+import daemonsetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/daemonsets/daemonsets-route.injectable";
 import workloadsSidebarItemInjectable from "../workloads/workloads-sidebar-item.injectable";
 import { sidebarItemInjectionToken } from "@k8slens/cluster-sidebar";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
-import navigateToStatefulsetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/statefulsets/navigate-to-statefulsets.injectable";
+import navigateToDaemonsetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/daemonsets/navigate-to-daemonsets.injectable";
 
-const statefulSetsSidebarItemInjectable = getInjectable({
-  id: "sidebar-item-stateful-sets",
+const daemonsetsSidebarItemInjectable = getInjectable({
+  id: "sidebar-item-daemonsets",
 
   instantiate: (di) => {
-    const route = di.inject(statefulsetsRouteInjectable);
+    const route = di.inject(daemonsetsRouteInjectable);
 
     return {
       parentId: workloadsSidebarItemInjectable.id,
-      title: "Stateful Sets",
-      onClick: di.inject(navigateToStatefulsetsInjectable),
+      title: "DaemonSets",
+      onClick: di.inject(navigateToDaemonsetsInjectable),
       isActive: di.inject(routeIsActiveInjectable, route),
       isVisible: route.isEnabled,
-      orderNumber: 50,
+      orderNumber: 40,
     };
   },
 
   injectionToken: sidebarItemInjectionToken,
 });
 
-export default statefulSetsSidebarItemInjectable;
+export default daemonsetsSidebarItemInjectable;
