+++ packages/core/src/renderer/components/workloads-replicasets/replicasets-sidebar-item.injectable.tsx
@@ -4,29 +4,29 @@
  */
 import { getInjectable } from "@ogre-tools/injectable";
 
-import daemonsetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/daemonsets/daemonsets-route.injectable";
+import replicasetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/replicasets/replicasets-route.injectable";
 import workloadsSidebarItemInjectable from "../workloads/workloads-sidebar-item.injectable";
 import { sidebarItemInjectionToken } from "@k8slens/cluster-sidebar";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
-import navigateToDaemonsetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/daemonsets/navigate-to-daemonsets.injectable";
+import navigateToReplicasetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/replicasets/navigate-to-replicasets.injectable";
 
-const daemonSetsSidebarItemInjectable = getInjectable({
-  id: "sidebar-item-daemon-sets",
+const replicasetsSidebarItemInjectable = getInjectable({
+  id: "sidebar-item-replicasets",
 
   instantiate: (di) => {
-    const route = di.inject(daemonsetsRouteInjectable);
+    const route = di.inject(replicasetsRouteInjectable);
 
     return {
       parentId: workloadsSidebarItemInjectable.id,
-      title: "Daemon Sets",
-      onClick: di.inject(navigateToDaemonsetsInjectable),
+      title: "ReplicaSets",
+      onClick: di.inject(navigateToReplicasetsInjectable),
       isActive: di.inject(routeIsActiveInjectable, route),
       isVisible: route.isEnabled,
-      orderNumber: 40,
+      orderNumber: 60,
     };
   },
 
   injectionToken: sidebarItemInjectionToken,
 });
 
-export default daemonSetsSidebarItemInjectable;
+export default replicasetsSidebarItemInjectable;
