+++ packages/core/src/renderer/components/workloads-statefulsets/statefulsets-sidebar-item.injectable.tsx
@@ -4,29 +4,29 @@
  */
 import { getInjectable } from "@ogre-tools/injectable";
 
-import replicasetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/replicasets/replicasets-route.injectable";
+import statefulsetsRouteInjectable from "../../../common/front-end-routing/routes/cluster/workloads/statefulsets/statefulsets-route.injectable";
 import workloadsSidebarItemInjectable from "../workloads/workloads-sidebar-item.injectable";
 import { sidebarItemInjectionToken } from "@k8slens/cluster-sidebar";
 import routeIsActiveInjectable from "../../routes/route-is-active.injectable";
-import navigateToReplicasetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/replicasets/navigate-to-replicasets.injectable";
+import navigateToStatefulsetsInjectable from "../../../common/front-end-routing/routes/cluster/workloads/statefulsets/navigate-to-statefulsets.injectable";
 
-const replicaSetsSidebarItemInjectable = getInjectable({
-  id: "sidebar-item-replica-sets",
+const statefulsetsSidebarItemInjectable = getInjectable({
+  id: "sidebar-item-statefulsets",
 
   instantiate: (di) => {
-    const route = di.inject(replicasetsRouteInjectable);
+    const route = di.inject(statefulsetsRouteInjectable);
 
     return {
       parentId: workloadsSidebarItemInjectable.id,
-      title: "Replica Sets",
-      onClick: di.inject(navigateToReplicasetsInjectable),
+      title: "StatefulSets",
+      onClick: di.inject(navigateToStatefulsetsInjectable),
       isActive: di.inject(routeIsActiveInjectable, route),
       isVisible: route.isEnabled,
-      orderNumber: 60,
+      orderNumber: 50,
     };
   },
 
   injectionToken: sidebarItemInjectionToken,
 });
 
-export default replicaSetsSidebarItemInjectable;
+export default statefulsetsSidebarItemInjectable;
